Pass section id to included page edit button, fix warning #245
This fixes another undefined index 'hid' warning. Further, if the
include plugin sets its own section id, this section id is passed to
the editor now. This allows jumping back to the place from where the
page has been included. Currently, these section ids are only set when
a page is included for the first time.